Add `serve-http` command
Add this command, which spawns an HTTP server which serves only a very
simple API endpoint for now, `/api/products`.

FastAPI is used as the framework to handle HTTP requests and responses,
dispatching the the handlers written in http.py.  Its external interface
is the `app` global in http.py, which implements the ASGI interface.

Uvicorn is used as the ASGI gateway.  It listens on the specified port
and handles the incoming HTTP traffic, forwarding requests to the
aforementioned `http.app`.

Add those two packages to pyproject.toml.  Also add python-dotenv, which
is an optional dependency of Uvicorn.  We need it to use the env_file
feature.

Use the `uvicorn.run` form where we pass the path to the `app` as a
string.  This runs the HTTP server in a new process under the hood,
which imports http.py, so that allows us to use the auto-reload feature.
We could also import `canadiantracker.http.app` ourselves and pass the
object to `uvicorn.run`, but then we wouldn't be able to use
auto-reload.

We need to pass the DB path to the HTTP server app for it to fetch the
data on products.  But since our "app" is imported in a separate
process, we can't pass it parameters directly.  Make the parent process
write a temporary env file and pass it using Uvicorn env_file parameter.
I think this is good since it allows http.py to be used in a standalone
way as well:

  CTSCRAPER_HTTP_DB_PATH=inventory.db uvicorn canadiantracker.http:app

This will be useful when we will want to deploy the web app in
production, behind an Apache server or whatever.  In that case, I think
it's typical for configurations for the particular appliation instance
to be passed through environment.

The endpoint added is more an example than anything, it returns a JSON
list of all products in the database (the conversion from Python data
structures to JSON is done implicitly by FastAPI I think).

This can be tested with:

    $ poetry run ctscraper serve-http --db-path inventory.db
    ...

    $ curl -s localhost:5000/api/products | python3 -m json.tool | head
    [
        {
            "name": "iRobot Roomba s9+ Wi Fi Connected Robot Vacuum Cleaner",
            "code": "0439415P"
        },
        {
            "name": "Dyson V11 Outsize Cordless Stick Vacuum Cleaner",
            "code": "0438495P"
        },
        {