Use appropriate key name for level
I don't know if it matters, but we use the search key "ast-id-level-1"
for all product searches by category.  If you browse here, for example:

    https://www.canadiantire.ca/en/kitchen/cookware/fry-pans.html?adlocation=LOL_DLP_Living_Kitchen_Cookware_FryPans_en

and look at the API requests done, you'll see:

    https://api.canadiantire.ca/search/api/v0/product/en/?site=ct%3Bstore%3D0064%3Bformat%3Djson%3Bcount%3D32%3Bx1%3Dast-id-level-3%3Bq1%3Dkitchen%3A%3Acookware%3A%3Afry-pans%3Bq%3D*%3Bcallback%3Dcallback

We see that the key is ast-id-level-3, which matches the number of
levels in the category kitchen::cookware::fry-pans.

Change the code to use the correct key name, based on the number of
levels in the category.  I didn't notice any changes in behaviors
though.