Add support for database migrations using Alembic
Add an Alembic setup and one revision file.  This will allow us to more
easily support database migration in the future.

After this commit, users will need to create a new database using the
instructions provided in the README.  That new database will contain an
`alembic_version` table that identifies the revision on which this
database is.  Users can migrate their data manually from their old
database to the new database relatively easily, since the schemas for
the application tables are the same.

Hopefully, thanks to migrations, this is the only time we ask users to
start from a fresh database.

The application no longer automatically creates a database file if the
specified file doesn't exist.  This is because we want to create it
using Alembic, so that it contains that `alembic_version` table.  It
might be possible to have the application create the database and stamp
it with the right Alembic revision hash, as described here:

    https://alembic.sqlalchemy.org/en/latest/cookbook.html#building-an-up-to-date-database-from-scratch

... but it's simpler for now to not do it.

The way Alembic works is that each revision file points to the one that
comes before it, a bit like a git commit points to its parent.  Each
revision has upgrade and downgrade functions that contain code to apply
or un-apply database modification.

The initial revision file was intially created using:

    $ CTRACKER_DB_PATH=inventory.db poetry run alembic revision --autogenerate -m "create initial tables"

Where inventory.db does not exist.  This makes the auto-generation
mechanism see that no tables exist, it knows that some tables should
exist, so it generates some code to create those tables.  The same
technique can be used for schema modifications later, for subsequent
revisions (with limitations, see the Alembic doc).

Otherwise, the initial content of the alembic directory and the alembic
.ini file is generated with `alembic init`.

The only modification done to alembic.ini is the note added before
`sqlalchemy.url`.  The modifications done to alembic/env.py are:

 - the code to set `sqlalchemy.url` dynamically based on an environment
   variable.  I don't really know how we can do otherwise, it's not like
   we can hard-code a path to the database file in alembic.ini: each
   user can store its database where they want, so there needs to be a
   way to specify it.  It's inspired by:

     https://stackoverflow.com/questions/59109682/sqlalchemy-sqlite-url-relative-to-home-or-environment-variable

 - importing sqlalchemy_base from canadiantracker.storage, which allows
   to populate the target_metadata variable.  This allows Alembic to
   know the current schema expected by the application, to make
   autogeneration to work.  The code in storage.py remains the source of
   truth for that schema.