Add ctquery to query the inventory from the command line
Add the skeleton of a new tool, ctquery, which allows users to query the
database from the command line. For the moment, the only supported
functionality allows a user to query the price history for a given
product code.

  $ poetry run ctquery price-history --db-path ~/CanadianTracker.db 0052133P
    [{'datetime': '2022-03-17T16:48:55.159020', 'price': '99.99'}, {'datetime': '2022-03-18T11:04:42.074893', 'price': '99.99'}]

Two output formats are supports: json and "plot".

The json format outputs an array of dictionnaries of the form:
  {"datetime": "2022-03-22T00:18:12.659088", "price": "579.99"}

The "plot" output format produces a plot on the terminal. It
is currently fairly crude, but functional.

Note that this results in a nagging warning about Decimal object
precision. I have not looked into this.

/home/jgalar/Projects/CanadianTracker/src/canadiantracker/storage.py:100: SAWarning: Dialect sqlite+pysqlite does *not* support Decimal objects natively, and SQLAlchemy must convert from floating point - rounding errors and other issues may occur. Please consider storing Decimal numbers as strings or integers on this platform for lossless storage.
  return result.all() if result else None

Signed-off-by: Jérémie Galarneau <jeremie.galarneau@gmail.com>