Add `ctserver server` command
Add this command, which spawns an HTTP server which serves a few
rudimentary endpoints:

 - /products: An HTML page with the list of all products
 - /products/$id: An HTML page with the information about product $id,
   including a chart with the price history
 - /api/products: JSON array with the list of all products
 - /api/products/$id/samples: JSON array with the list of price samples
   for product $id.

FastAPI is used as the framework to handle HTTP requests and responses,
dispatching the the handlers written in http.py.  Its external interface
is the `app` global in http.py, which implements the ASGI interface.

Uvicorn is used as the ASGI gateway.  It listens on the specified port
and handles the incoming HTTP traffic, forwarding requests to the
aforementioned `http.app`.

Add those two packages to pyproject.toml.  Also add python-dotenv, which
is an optional dependency of Uvicorn.  We need it to use the env_file
feature, to pass the database path down to the subprocess (described
below).

Use the `uvicorn.run` form where we pass the path to the `app` as a
string.  This runs the HTTP server in a new process under the hood,
which imports http.py.  This allows us to use the auto-reload feature.
We could also import `canadiantracker.http.app` ourselves and pass the
object to `uvicorn.run`, but then we wouldn't be able to use
auto-reload.

We need to pass the DB path to the HTTP server app for it to fetch the
data on products.  But since our "app" is imported in a separate
process, we can't pass it parameters directly.  Make the parent process
write a temporary env file and pass it using Uvicorn env_file parameter.
I think this is good since it allows http.py to be used in a standalone
way as well:

  CTSERVER_SERVE_DB_PATH=inventory.db uvicorn canadiantracker.http:app

This will be useful when we will want to deploy the web app in
production, behind an Apache server or whatever.  In that case, I think
it's typical for configurations for the particular appliation instance
to be passed through environment.

This can be used with:

    $ poetry run ctserver serve --db-path inventory.db
    ...

You can then navigate to http://127.0.0.1:5000/.