Record URL in products_static table
It would be useful to have the URL to the product to be able to provide
links from third party tools to the Canadian Tire website.  This patch
makes ctscraper record the url (e.g.
`/en/pdp/antiskid-sand-10-kg-0597113p.html`) in the products_static
table.

Provide a new migration script to add a `url` column in that table.  The
migration can be applied by running

  $ CTRACKER_DB_PATH=test.db poetry run alembic upgrade head

The migration must be applied before attempting to run the code in this
patch.

When scraping a product and an entry already exists, check whether the
url property is None (NULL in SQL), and if so update it.