Error out if specifying a database file that doesn't exist
Specifying a database file (using --db-path) that doesn't exist causes
an empty sqlite database to be created.  The program then fails, because
the expected tables don't exist in the database.  Try to avoid this by
verifying if the specified file exists before trying to open a
connection to it.  This is not perfect, as it's a TOCTOU bug, but
realistically it's good enough for this program, at least for now.