Record sku in products_static table
Product sku numbers (also called diplay product numbers in other parts
of the API) are user visible product codes. They are presented in
stores, on receipts, and on the CT website.

Storing the SKU will allow us to display product codes in a more
intuitive way to Domain Experts.

SKUs are apparently derived from the product code, but they contain an
additional number whose meaning is not entirely clear.

A new migration script is added to add a `sku` column to
`products_static`. The migration can be applied by running:

  $ CTRACKER_DB_PATH=test.db poetry run alembic upgrade head

The migration must be applied before attempting to run the code in
this patch.

When scraping a product and an entry already exists, check whether
the sku property is None (NULL in SQL) and, if so, update it.

Signed-off-by: Jérémie Galarneau <jeremie.galarneau@gmail.com>