storage: verify alembic revision id when opening database
On of my greatest fears is to run the Canadian Tracker code with a
database that is at the wrong migration, possibly corrupting the data.
To avoid this, I propose to add a check, when opening the database, that
the database is at the right revision.

This also helps for the case of specifying a file that doesn't exist as
the database.  We recently added a file existence check for that
recently, this patch makes it not really needed anymore.  With this
patch, when specifying a file that doesn't exist, an empty sqlite
database will be created, but the program will error out with:

    Database is missing the alembic_version table. Make sure it is initialized correctly.

The user can then run the necessary "alembic upgrade head" command to
populate the database.

When the revision is not the right one, the message is:

    Failed to validate database revision: expected 8fe398b58b5f, got ac4b7897c153

    Make sure the database is at the latest revision.