Handle sku migrations from one product entry to another
When listing the product inventory, the triangle API can return a sku
that was previously associated to a different product entry.

These kind of "migrations" of skus happen periodically when a sku is
re-parented to a different product (typically because it changed names).
In this case, edit the existing sku entry to point to the more recent
product listing entry.

Signed-off-by: Jérémie Galarneau <jeremie.galarneau@gmail.com>