Update name and URL when trying to add existing product
Product names and URLs are susceptible to change, notably when a
multi-variant product gets split and the original product becomes a
specific variant.  Update them when trying to add a product that already
exists.

Update the is_in_clearance flag as well, since we want to know when a
product becomes in clearance.