Convert Javascript code to Typescript
Move the Javascript code to some .ts files.  Add some tsc / webpack
config files that I don't fully understand so that we generate one .js
per page (index and sku) that bundle all the dependencies.

Before this patch, we would rely on external CDNs for third party
libraries, whereas now everything is bundled as part of the build
process.  Speaking of build process, it is documented in the README.

We need the explicit include in pyproject.toml, because the dist
directory is ignored by .gitignore, and so would not be automatically
included by `poetry build`.

No change in behavior is expected with this patch.