Add `prune-samples` command
Since the production database size is growing out of control, I suggest
adding a way to get rid of past redundant samples, where "redundant"
means that a sample has the same price value as the previous sample for
the same SKU, and is not the very last sample for that SKU.  The
downside of this is that we lose some information about when a product
price was sampled, we only keep some intervals.  However, I claim it's
not very important given the main goal of the project.

The last snapshot I have of the database is 19GB.  After running
"prune-samples" and VACUUM (the sqlite comment) on it, it is down to
about 800M.  The database will continue to grow even if keeping only the
non-redundant samples, but it will grow at a much slower pace.