Add --discard-equal option to scrape-prices
This is the counter-part to the new prune-samples command.
prune-samples can be used to remove redundant samples from an existing
database.  The new --discard-equal option to scrape-prices makes it so
that when adding a new sample, the previous last sample is deleted if it
has the same price as the new sample.  This keeps the database free of
redundant samples in the first place.

I made it an option in order to accomodate the different use cases of
the numerous CanadianTracker users, but it could also be made the
default behavior, with or without an option.