Change type of samples price column to integer
When querying the database from Python/SQLAlchemy, we get this warning:

    $ poetry run ctquery price-history --db-path inventory.db 027-8158-6
    /home/simark/src/CanadianTracker/src/canadiantracker/query.py:123: SAWarning: Dialect sqlite+pysqlite does *not* support Decimal objects natively, and SQLAlchemy must convert from floating point - rounding errors and other issues may occur. Please consider storing Decimal numbers as strings or integers on this platform for lossless storage.

We can avoid this and all the problems related to floating point /
decimal numbers by storing the price as an integer representing the
number of cents.

Add a migration for the database to introduce a new "price_cents"
integer column, and remove the existing "price" column.

Change _StorageProductSample to transparently convert to and from this
price_cents column:

 - in the constructor, accept a Decimal value in dollars, but transform
   it in cents when assigning to the price_cents field
 - add a price property that converts the raw price_cents value to
   dollars, which the rest of the code can use transparently

While at it, change how the json parser parses decimal number, when
getting data from the CT API, so that we get Decimal objects instead of
float objects (the change in triangle.py).  Adjust ProductInfo.price in
model.py accordingly.  This way, we never deal with floats in the whole
pipeline.