Bump pytest version to get rid of vulnerability warning
Dependabot warns that:

The py library through 1.11.0 for Python allows remote attackers to
conduct a ReDoS (Regular expression Denial of Service) attack via a
Subversion repository with crafted info data, because the InfoSvnCommand
argument is mishandled.

The particular codepath in question is the regular expression at
py._path.svnurl.InfoSvnCommand.lspattern and is only relevant when
dealing with subversion (svn) projects. Notablely the codepath is not
used in the popular pytest project. The developers of the pytest package
have released version 7.2.0 which removes their dependency on py. Users
of pytest seeing alerts relating to this advisory may update to version
7.2.0 of pytest to resolve this issue. See pytest-dev/py#287 (comment)
for additional context.

Signed-off-by: Jérémie Galarneau <jeremie.galarneau@gmail.com>