Make ProductRepository.products a function, add filter
The following patch needs to be able to filter the list of products
obtained by ProductRepository.products.  And since it would be nice to
be able to filter at the database level (rather than having the caller
filter at the Python level), the caller needs to be able to pass some
filter data.  Turn the products property into a method, and add a
parameter that allows filtering on product code.