Standardize intra-project imports
This is subjective / personal preference.  I like to import modules
instead of types / functions directly, and then refer to the imported
types / functions using their module name.  For instance, instead of
doing:

    import canadiantracker.model

    ... use canadiantracker.model.ProductListingEntry ...

or:

    from canadiantracker.model import ProductListingEntry

    ... use ProductListingEntry ...

I prefer:

    from canadiantracker import model

    ... use model.ProductListingEntry ...

The first version is unnecessarily too verbose.  The second version does
not make it clear (when reading the code, far from the import) that this
is an imported type, and where it comes from.  I think that the third
version strikes a good balance.

Change all the intra-project imports such that we refer to imported
types and functions using the module name.