Add Github ACtions CI configuration
Add a Github Actions workflow to run some checks when a branch is pushed
to and when a pull request is created.  The job installs the
dependencies, runs black to verify the code is properly formatted, and
runs `ctscraper --help`.  Later, this last step will be replaced by
proper tests.