Validate product code format in ProductRepository API
We see product codes with both upper and lower case `p`.  We currently
normalize it to upper case when inserting products in the database.
However, we don't normalize it in other API calls, leading to weird
behaviors like:

    repository.add_product(model.Product("1234567p", "Hello", False, "/foo"))
    product = repository.get_product_by_code("1234567p")
    # product is None here :(

The most common formatting I could find was with an upper case P.  So I
propose that the storage API enforces this formatting at its entry
points.