Make triangle use its own data types instead of model
The data returned by the "triangle" API does not necessarily match that
model.  For instance, the model will say that products have an attribute
that is a list of SKU model objects.  The products returned by the
triangle API does not have that.  Make it return its own data objects,
which are basically a Python representation of what the Canadian Tire
API returns.

Also, a subsequent commit removes the model objects entirely.