scrape-prices: keep going when a price query fails
Running scrape-prices on the production database, I see a batch of price
queries getting some error 400, which aborts the whole run.  It's not
clear why this particular batch fails.  I tried re-running the same
batch, varying batch sizes and removing some SKUs, and I could not find
a clear culprit.

In this case, it would be more helpful to just keep going with the price
scraping.  We won't get any prices for the SKUs in the failing batch,
but it's better than getting no prices at all.

When getting the specific error I encountered (see below), skip the
retry, since it's just wasted time.  Then, in __iter__, swallow the
exception (we'll have logged some things already) and keep going.
Define and use a more specific exception type, to avoid catching and
silencing other unrelated errors, which we might need to handle
differently.

For reference, the error I get is:

    ERROR:canadiantracker.triangle:Got status code 400 on try 0
    ERROR:canadiantracker.triangle:{
      "statusCode": 400,
      "errors": [
        "Hybris error: Error: There is a problem with hybris product details Service call: Request failed with status code 404"
      ],
      "isCdsError": true,
      "error": "Bad Request",
      "errCode": "0",
      "xRequestId": null,
      "requestId": "e30f0882125de43d39bb4a6069291b48",
      "correlationId": "6359b575-2e1f-4f79-af51-14666b74b886"
    }